pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chardet>=5.2.0",
    "openai>=1.82.1",
    "pillow>=11.2.1",
    "python-magic>=0.4.27",
]
